+++ src/main/java/losty/netatmo/oauthtoken/OAuthTokenStore.java
@@ -1,4 +1,4 @@
-package losty.netatmo.store;
+package losty.netatmo.oauthtoken;
 
 /**
  * Holds the OAuth access token. May persist the token additionally.
